"check_style_score -> 5 of 5 passed. style score: 0/1 "

+++ cspp2- assignments/m2/Assignment-3/Solution.java
@@ -17,7 +17,7 @@ public class Solution {
 		Scanner s = new Scanner(System.in);      
         int base = s.nextInt();
         int exponent = s.nextInt();
-        double result = power(base,exponent);
+        long result = power(base,exponent);
         System.out.println(result);
 	}
 	/**
@@ -28,7 +28,7 @@ public class Solution {
 	 *
 	 * @return     { description_of_the_return_value }
 	 */
-	static double power(int base, int exponent) {
+	static long power(int base, int exponent) {
 	if(exponent == 0)
 	{
 		return 1;
